# api/openapi.yaml
openapi: 3.1.0
info:
  title: Speccy API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
paths:
...
  /docs/{docId}:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: getSpecDoc
      summary: Get one spec doc.
      responses:
        "200":
          description: The spec doc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDoc"
        default:
          $ref: "#/components/responses/Problem"
...
  /docs/{docId}/runs:
    parameters:
      - $ref: "#/components/parameters/DocId"
    post:
      operationId: startRun
      summary: Start a full review of the current version (REQ-020). Lint runs on its own on every save.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "202":
          description: The queued run. Follow its progress at /runs/{runId}/events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Problem"
    get:
      operationId: listRuns
      summary: List the review runs of a bundle, newest first.
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: The runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
        default:
          $ref: "#/components/responses/Problem"
...
  /docs/{docId}/runs/estimate:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: estimateRun
      summary: Estimate the tokens and cost of a full review before it starts (REQ-104).
      responses:
        "200":
          description: The estimate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEstimate"
        default:
          $ref: "#/components/responses/Problem"
  /docs/{docId}/assumptions:
    parameters:
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: listAssumptions
      summary: List the sentences of the current main doc that start with "Assumption:" (REQ-033).
      responses:
        "200":
          description: The assumptions.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Anchor"
        default:
          $ref: "#/components/responses/Problem"
...
  /runs/{runId}/events:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: runEvents
      summary: Follow a run's progress by stage, as server-sent events (REQ-026).
      responses:
        "200":
          description: A stream of events. Each data line is a RunEvent. The stream ends when the run ends.
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
  /runs/{runId}/claims:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: listClaims
      summary: List a run's factual claims and their labels (REQ-031).
      responses:
        "200":
          description: The claims, in document order.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Claim"
        default:
          $ref: "#/components/responses/Problem"
  /runs/{runId}/questions:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: listQuestions
      summary: List a run's build questions, reader answers, and results (REQ-040 to REQ-046).
      responses:
        "200":
          description: The questions in order. Empty for a run without a divergence test.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BuildQuestion"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DocId:
      name: docId
      in: path
      required: true
      description: The ID of one spec doc.
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    Problem:
      description: An error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details with a stable code.
      required: [type, title, status, code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: A stable machine-readable error code.
    SpecDoc:
      type: object
      description: One spec doc of a bundle, with its own profile, versions, review runs and verdict.
      required: [id, bundle_id, slug, title, profile_key, source_kind, path, current_version, updated_at]
      properties:
        id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
          description: The bundle that holds the spec doc.
        slug:
          type: string
        title:
          type: string
        profile_key:
          type: string
          description: The profile of the spec doc.
        source_kind:
          type: string
          enum: [local, db, github]
        github:
          $ref: "#/components/schemas/BundleGithub"
        path:
          type: string
          description: The path of the spec doc in its bundle.
        current_version:
          $ref: "#/components/schemas/Version"
        updated_at:
          type: string
          format: date-time
        verdict:
          $ref: "#/components/schemas/BundleVerdict"
        run_error:
          type: string
          description: Why the latest run on the current version failed, when it failed.
        adopt:
          $ref: "#/components/schemas/Adopt"
        status:
          $ref: "#/components/schemas/ReviewStatus"
        next_action:
          $ref: "#/components/schemas/NextAction"
...
    NextAction:
      type: object
      description: The one thing the caller must do next on this bundle. Absent when nothing is open. The list carries the kind and the sentence; one bundle also carries the target.
      required: [kind, sentence]
      properties:
        kind:
          type: string
          enum: [waiver, decide, fix, review, adopt, request_review, handoff]
        sentence:
          type: string
          description: What to do, in words, for a button label or a status line.
        waiver_id:
          type: string
          format: uuid
        finding_id:
          type: string
          format: uuid
        tour_key:
          type: string
          description: The key of the tour point to open.
    BundleVerdict:
      type: object
      description: The verdict of the bundle's latest completed run. It is stale when that run is not on the current version.
      required: [run_id, version_number, result, score, radar, waiver_count, relaxed_count, must, should, info, kind]
      properties:
        run_id:
          type: string
          format: uuid
        version_number:
          type: integer
          format: int64
        kind:
          type: string
          enum: [lint, full]
          description: lint means only the lint stage ran.
        result:
          $ref: "#/components/schemas/VerdictResult"
        score:
          type: integer
        radar:
          type: object
          additionalProperties:
            type: integer
        waiver_count:
          type: integer
        relaxed_count:
          type: integer
          description: Checks in adoption mode (REQ-133).
        must:
          type: integer
          description: Open MUST findings.
        should:
          type: integer
        info:
          type: integer
        blocking_threads:
          type: integer
          description: Open blocking threads. Any makes the verdict Not Build Ready (§8.6 rule 2).
        ai_run_id:
          type: string
          format: uuid
          description: The full review whose AI findings this verdict counts. For a full run, the run itself.
        ai_version_number:
          type: integer
          format: int64
          description: The version the AI review read, when it is older than this verdict's version.
        sections_changed:
          type: integer
          description: How many sections changed since the AI review read the doc.
        stale_reason:
          type: string
          enum: [upstream_changed]
          description: Set when the verdict is stale because a linked bundle has a newer version than the run read (REQ-056).
    VerdictResult:
      type: string
      enum: [build_ready, not_build_ready, stale]
...
    Run:
      type: object
      required: [id, doc_id, version_id, version_number, profile_key, profile_version, kind, status, stage, error, started_at]
      properties:
        id:
          type: string
          format: uuid
        doc_id:
          type: string
          format: uuid
        version_id:
          type: string
          format: uuid
        version_number:
          type: integer
          format: int64
        profile_key:
          type: string
        profile_version:
          type: integer
          format: int64
        kind:
          type: string
          enum: [lint, full]
        status:
          type: string
          enum: [queued, running, complete, failed]
        stage:
          type: string
        error:
          type: string
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        tokens_in:
          type: integer
          format: int64
        tokens_out:
          type: integer
          format: int64
        cost_estimate:
          type: number
          description: USD, from the prices on the role assignments.
        cache_hits:
          type: integer
          format: int64
          description: Steps answered from the cache (REQ-021).
        notes:
          type: array
          description: Notes for the run report, such as "no search source configured" (REQ-034).
          items:
            type: string
        verdict:
          $ref: "#/components/schemas/BundleVerdict"
    RunEstimate:
      type: object
      required: [calls, cached_steps, tokens_in, tokens_out, priced]
      properties:
        calls:
          type: integer
        cached_steps:
          type: integer
        tokens_in:
          type: integer
          format: int64
        tokens_out:
          type: integer
          format: int64
        cost_usd:
          type: number
        priced:
          type: boolean
          description: False when the reviewer role has no prices, so there is no cost.
...
    ClaimSource:
      type: object
      description: One source of a claim, with what the metadata resolver learned about it.
      required: [url]
      properties:
        url:
          type: string
          description: The address the search returned.
        final_url:
          type: string
          description: The address that answered, after the redirects. Empty when the resolver did not run.
        chain:
          type: array
          description: Every address Speccy requested, in order.
          items:
            type: string
        status:
          type: integer
        retrieved_at:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        tier:
          type: string
          enum: [primary, secondary]
        dropped:
          type: boolean
          description: The source policy refused the source.
        reason:
          type: string
    Claim:
      type: object
      required: [id, text, label, reason, sources, anchor, class]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        label:
          type: string
          enum: [verified, contradicted, unverified]
        reason:
          type: string
        class:
          type: string
          description: The claim class the profile gave the section, or unclassified.
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ClaimSource"
        anchor:
          $ref: "#/components/schemas/Anchor"
...
    BuildQuestion:
      type: object
      required: [id, number, text, level, cites, anchor, result, answers, groups]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        text:
          type: string
        level:
          type: string
          enum: [MUST, SHOULD]
        cites:
          type: array
          items:
            $ref: "#/components/schemas/Cite"
        anchor:
          $ref: "#/components/schemas/Anchor"
        result:
          type: string
          enum: [agree, diverge, gap]
        answers:
          type: array
          description: One per reader, in reader order. Readers are named by number only (DEC-013).
          items:
            $ref: "#/components/schemas/ReaderAnswer"
        groups:
          type: array
          description: The judge's groups of reader numbers with the same meaning.
          items:
            type: array
            items:
              type: integer
    Cite:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [section, trace]
        path:
          type: array
          items:
            type: string
        id:
          type: string
    ReaderAnswer:
      type: object
      required: [reader, answer, answered, quotes]
      properties:
        reader:
          type: integer
        answer:
          type: string
        answered:
          type: boolean
          description: False for NOT SPECIFIED, and for an answer with no quote found in the bundle (REQ-043).
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/QuoteCheck"
    QuoteCheck:
      type: object
      required: [text, found]
      properties:
        text:
          type: string
        found:
          type: boolean
...
    RunList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Run"
    Anchor:
      type: object
      description: A range of text with context (SDD §8.8).
      required: [file, heading_path, quote, prefix, suffix, start, end]
      properties:
        file:
          type: string
        heading_path:
          type: array
          items:
            type: string
        quote:
          type: string
        prefix:
          type: string
        suffix:
          type: string
        start:
          type: integer
          description: Byte offset of the quote in the file.
        end:
          type: integer
        detached:
          type: boolean
          description: The text changed, and Speccy cannot find the quote in the current version (SDD §8.8).
...
    StartRunRequest:
      type: object
      properties:
        stages:
          type: array
          description: The model stages to run. Absent means all. Lint and the verdict always run.
          items:
            type: string
            enum: [rubric, grounding, divergence, coherence]
...
    BundleGithub:
      type: object
      description: Where a GitHub bundle comes from, and its draft (REQ-123).
      required: [repo, branch, path, draft, ahead]
      properties:
        repo:
          type: string
        branch:
          type: string
        path:
          type: string
        draft:
          type: boolean
          description: The current version has changes that are not on GitHub.
        ahead:
          type: boolean
          description: GitHub changed after the draft started.
        pr_url:
          type: string
...
    Adopt:
      type: object
      description: The frontmatter keys the main doc does not name, and the values a review used for them (REQ-135). Absent when the doc names both.
      properties:
        type:
          type: string
        size:
          type: string
...
    ReviewStatus:
      type: string
      enum: [draft, in_review, approved, superseded]
...
    Version:
      type: object
      required: [id, number, created_by, message, created_at]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
          format: int64
        created_by:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
...